allowed spawn_cleanup to be unregistered
using esky in a launcher process that controls sub processes. the launcher can relaunch the sub processes without needing to restart itself. this in turn means that auto_update can get called more than once (each time registers another cleanup). 

this change moves the spawn_cleanup code out, so that we can register & unregister it, and it uses that to keep the cleanup down to just one call. (the latest call that should be run with the up to date version)